+++ cartographer/mapping/internal/optimization/cost_functions/cost_helpers_impl.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_HELPERS_IMPL_H_
-#define CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_HELPERS_IMPL_H_
+#ifndef CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_COST_HELPERS_IMPL_H_
+#define CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_COST_HELPERS_IMPL_H_
 
 #include "Eigen/Core"
 #include "cartographer/transform/rigid_transform.h"
@@ -194,4 +194,4 @@ InterpolateNodes2D(const T* const prev_node_pose,
 }  // namespace mapping
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_HELPERS_IMPL_H_
+#endif  // CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_COST_HELPERS_IMPL_H_
