+++ cartographer/mapping/internal/optimization/cost_functions/landmark_cost_function_2d.h
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_LANDMARK_COST_FUNCTION_3D_H_
-#define CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_LANDMARK_COST_FUNCTION_3D_H_
+#ifndef CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_LANDMARK_COST_FUNCTION_2D_H_
+#define CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_LANDMARK_COST_FUNCTION_2D_H_
 
 #include "Eigen/Core"
 #include "Eigen/Geometry"
-#include "cartographer/mapping/internal/optimization/cost_helpers.h"
-#include "cartographer/mapping/internal/optimization/optimization_problem_3d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/cost_helpers.h"
+#include "cartographer/mapping/internal/optimization/optimization_problem_2d.h"
 #include "cartographer/mapping/pose_graph_interface.h"
 #include "cartographer/transform/rigid_transform.h"
 #include "cartographer/transform/transform.h"
@@ -32,37 +32,32 @@ namespace mapping {
 namespace optimization {
 
 // Cost function measuring the weighted error between the observed pose given by
-// the landmark measurement and the linearly interpolated pose.
-class LandmarkCostFunction3D {
+// the landmark measurement and the linearly interpolated pose of embedded in 3D
+// space node poses.
+class LandmarkCostFunction2D {
  public:
   using LandmarkObservation =
       PoseGraphInterface::LandmarkNode::LandmarkObservation;
 
   static ceres::CostFunction* CreateAutoDiffCostFunction(
-      const LandmarkObservation& observation, const NodeSpec3D& prev_node,
-      const NodeSpec3D& next_node) {
+      const LandmarkObservation& observation, const NodeSpec2D& prev_node,
+      const NodeSpec2D& next_node) {
     return new ceres::AutoDiffCostFunction<
-        LandmarkCostFunction3D, 6 /* residuals */,
-        4 /* previous node rotation variables */,
-        3 /* previous node translation variables */,
-        4 /* next node rotation variables */,
-        3 /* next node translation variables */,
+        LandmarkCostFunction2D, 6 /* residuals */,
+        3 /* previous node pose variables */, 3 /* next node pose variables */,
         4 /* landmark rotation variables */,
         3 /* landmark translation variables */>(
-        new LandmarkCostFunction3D(observation, prev_node, next_node));
+        new LandmarkCostFunction2D(observation, prev_node, next_node));
   }
 
   template <typename T>
-  bool operator()(const T* const prev_node_rotation,
-                  const T* const prev_node_translation,
-                  const T* const next_node_rotation,
-                  const T* const next_node_translation,
+  bool operator()(const T* const prev_node_pose, const T* const next_node_pose,
                   const T* const landmark_rotation,
                   const T* const landmark_translation, T* const e) const {
     const std::tuple<std::array<T, 4>, std::array<T, 3>>
-        interpolated_rotation_and_translation = InterpolateNodes3D(
-            prev_node_rotation, prev_node_translation, next_node_rotation,
-            next_node_translation, interpolation_parameter_);
+        interpolated_rotation_and_translation = InterpolateNodes2D(
+            prev_node_pose, prev_node_gravity_alignment_, next_node_pose,
+            next_node_gravity_alignment_, interpolation_parameter_);
     const std::array<T, 6> error = ScaleError(
         ComputeUnscaledError(
             landmark_to_tracking_transform_,
@@ -75,11 +70,13 @@ class LandmarkCostFunction3D {
   }
 
  private:
-  LandmarkCostFunction3D(const LandmarkObservation& observation,
-                         const NodeSpec3D& prev_node,
-                         const NodeSpec3D& next_node)
+  LandmarkCostFunction2D(const LandmarkObservation& observation,
+                         const NodeSpec2D& prev_node,
+                         const NodeSpec2D& next_node)
       : landmark_to_tracking_transform_(
             observation.landmark_to_tracking_transform),
+        prev_node_gravity_alignment_(prev_node.gravity_alignment),
+        next_node_gravity_alignment_(next_node.gravity_alignment),
         translation_weight_(observation.translation_weight),
         rotation_weight_(observation.rotation_weight),
         interpolation_parameter_(
@@ -87,6 +84,8 @@ class LandmarkCostFunction3D {
             common::ToSeconds(next_node.time - prev_node.time)) {}
 
   const transform::Rigid3d landmark_to_tracking_transform_;
+  const Eigen::Quaterniond prev_node_gravity_alignment_;
+  const Eigen::Quaterniond next_node_gravity_alignment_;
   const double translation_weight_;
   const double rotation_weight_;
   const double interpolation_parameter_;
@@ -96,4 +95,4 @@ class LandmarkCostFunction3D {
 }  // namespace mapping
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_LANDMARK_COST_FUNCTION_3D_H_
+#endif  // CARTOGRAPHER_MAPPING_INTERNAL_OPTIMIZATION_COST_FUNCTIONS_LANDMARK_COST_FUNCTION_2D_H_
