+++ cartographer/mapping/internal/optimization/optimization_problem_2d.cc
@@ -28,9 +28,8 @@
 #include "cartographer/common/histogram.h"
 #include "cartographer/common/math.h"
 #include "cartographer/mapping/internal/optimization/ceres_pose.h"
-#include "cartographer/mapping/internal/optimization/cost_helpers.h"
-#include "cartographer/mapping/internal/optimization/landmark_cost_function_2d.h"
-#include "cartographer/mapping/internal/optimization/spa_cost_function_2d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/landmark_cost_function_2d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/spa_cost_function_2d.h"
 #include "cartographer/sensor/odometry_data.h"
 #include "cartographer/transform/transform.h"
 #include "ceres/ceres.h"
