+++ cartographer/mapping/internal/optimization/optimization_problem_3d.cc
@@ -30,13 +30,13 @@
 #include "cartographer/common/make_unique.h"
 #include "cartographer/common/math.h"
 #include "cartographer/common/time.h"
-#include "cartographer/mapping/internal/3d/acceleration_cost_function_3d.h"
 #include "cartographer/mapping/internal/3d/imu_integration.h"
-#include "cartographer/mapping/internal/3d/rotation_cost_function_3d.h"
 #include "cartographer/mapping/internal/3d/rotation_parameterization.h"
 #include "cartographer/mapping/internal/optimization/ceres_pose.h"
-#include "cartographer/mapping/internal/optimization/landmark_cost_function_3d.h"
-#include "cartographer/mapping/internal/optimization/spa_cost_function_3d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/acceleration_cost_function_3d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/landmark_cost_function_3d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/rotation_cost_function_3d.h"
+#include "cartographer/mapping/internal/optimization/cost_functions/spa_cost_function_3d.h"
 #include "cartographer/transform/timestamped_transform.h"
 #include "cartographer/transform/transform.h"
 #include "ceres/ceres.h"
@@ -425,7 +425,7 @@ void OptimizationProblem3D::Solve(
               trajectory_data.imu_calibration.data());
         }
         problem.AddResidualBlock(
-            RotationCostFunction::CreateAutoDiffCostFunction(
+            RotationCostFunction3D::CreateAutoDiffCostFunction(
                 options_.rotation_weight(), result.delta_rotation),
             nullptr /* loss function */, C_nodes.at(first_node_id).rotation(),
             C_nodes.at(second_node_id).rotation(),
